+++ CHANGELOG.md
@@ -7,10 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+_No unreleased changes._
+
+## [1.11.3] - 2023-02-26
+
 ### Deprecated
 
-- The `zmodule check-dumpfile` action. The [completion] module alone will handle checking the
-  dumpfile.
+- The `zmodule check-dumpfile` action. The [completion] module alone now handles
+  checking the dumpfile. (See
+  [zimfw/completion#12](https://github.com/zimfw/completion/pull/12))
 
 ## [1.11.2] - 2023-02-16
 
@@ -332,7 +337,8 @@ Take your time to review the updated [README.md] and the changes listed below.
 [termtitle]: https://github.com/zimfw/termtitle
 [s1ck94]: https://github.com/zimfw/s1ck94
 
-[Unreleased]: https://github.com/zimfw/zimfw/compare/v1.11.2...HEAD
+[Unreleased]: https://github.com/zimfw/zimfw/compare/v1.11.3...HEAD
+[1.11.3]: https://github.com/zimfw/zimfw/compare/v1.11.2...v1.11.3
 [1.11.2]: https://github.com/zimfw/zimfw/compare/v1.11.1...v1.11.2
 [1.11.1]: https://github.com/zimfw/zimfw/compare/v1.11.0...v1.11.1
 [1.11.0]: https://github.com/zimfw/zimfw/compare/v1.10.0...v1.11.0